store exam contents as jsonb instead of files
- add info and files jsonb fields to exam
- remove uploads activerecord model
- remove db/seeds in favor of loading fixtures
- move mark processing to separate file
- remove schema checker, json schema coming soon
