--- app/helpers/exam-schema.yaml
@@ -1,188 +1,0 @@
----
-start: Exam
-Exam:
-  object:
-    - key: "policies"
-      value: Policies
-      optional: true
-    - key: "versions"
-      value: Versions
-      minSize: 1
-Policies:
-  array: Policy
-Policy: [String, "ignore-lockdown|tolerate-windowed"]
-Versions:
-  array: Version
-Version:
-  object:
-    - key: "instructions"
-      value: HTML
-      optional: true
-    - key: "questions"
-      value: Questions
-      minSize: 1
-    - key: "reference"
-      value: FileDirs
-      optional: true
-Questions:
-  array: Question
-Question:
-  object:
-    - key: "name"
-      value: String
-      optional: true
-    - key: "description"
-      value: HTML
-      optional: true
-    - key: "separateSubparts"
-      value: Boolean
-      optional: true
-    - key: "parts"
-      value: Parts
-    - key: "reference"
-      value: FileDirs
-      optional: true
-Parts:
-  array: Part
-Part:
-  object:
-    - key: "name"
-      value: String
-      optional: true
-    - key: "description"
-      value: HTML
-      optional: true
-    - key: "points"
-      value: Float
-    - key: "body"
-      value: BodyItems
-    - key: "reference"
-      value: FileDirs
-      optional: true
-FileDirs:
-  array: FileDir
-FileDir:
-  one_of:
-    - File
-    - Dir
-File:
-  object:
-    - key: "file"
-      value: String
-Dir:
-  object:
-    - key: "dir"
-      value: String
-BodyItems:
-  array: BodyItem
-BodyItem:
-  one_of:
-    - HTML
-    - Response
-HTMLs:
-  array: HTML
-  minSize: 1
-Response:
-  tagged_one_of:
-    - Code
-    - YesNo
-    - TrueFalse
-    - MultipleChoice
-    - Matching
-    - AllThatApply
-    - CodeTag
-    - Text
-Code:
-  object:
-    - key: "prompt"
-      value: HTMLs
-      optional: true
-    - key: "lang"
-      value: String
-    - key: "initial"
-      value: String
-      optional: true
-YesNo:
-  one_of:
-    - Boolean
-    - YesNoPrompt
-YesNoPrompt:
-  object:
-    - key: "prompt"
-      value: HTMLs
-      optional: true
-    - key: "correctAnswer"
-      value: Boolean
-TrueFalse:
-  one_of:
-    - Boolean
-    - TrueFalsePrompt
-TrueFalsePrompt:
-  object:
-    - key: "prompt"
-      value: HTMLs
-      optional: true
-    - key: "correctAnswer"
-      value: Boolean
-MultipleChoice:
-  object:
-    - key: "prompt"
-      value: HTMLs
-      optional: true
-    - key: "options"
-      value: Options
-    - key: "correctAnswer"
-      value: [Integer, [0, "options"]]
-Options:
-  array: HTML
-Matching:
-  object:
-    - key: "prompt"
-      value: HTMLs
-      optional: true
-    - key: "prompts"
-      value: Options
-    - key: "values"
-      value: Options
-    - key: "correctAnswers"
-      value: MatchingAnswers
-MatchingAnswers:
-  array: Integer
-AllThatApply:
-  object:
-    - key: "prompt"
-      value: HTMLs
-      optional: true
-    - key: "options"
-      value: ATAOptions
-ATAOptions:
-  array: ATAOption
-ATAOption:
-  dictionary:
-    key: HTML
-    value: Boolean
-    size: 1
-CodeTag:
-  object:
-    - key: "prompt"
-      value: HTMLs
-      optional: true
-    - key: "correctAnswer"
-      value: TagAnswer
-    - key: "choices"
-      value: [String, "all|question|part"]
-TagAnswer:
-  object:
-    - key: "filename"
-      value: String
-    - key: "line"
-      value: Integer
-Text:
-  one_of:
-    - Nil
-    - TextPrompt
-TextPrompt:
-  object:
-    - key: "prompt"
-      value: HTMLs
-
