+++ app/models/room.rb
@@ -2,6 +2,8 @@ class Room < ApplicationRecord
   belongs_to :exam
   has_many :registrations
 
+  validates_presence_of :name
+
   def finalized?
     registrations.all?(&:final)
   end
