+++ app/views/exams/new.html.erb
@@ -3,6 +3,6 @@
 <%= bootstrap_form_for(@exam) do |f| %>
   <%= f.text_field :name %>
   <%= f.check_box :enabled %>
-  <%= f.file_field :yaml %>
+  <%= f.file_field :file %>
   <%= f.submit %>
 <% end %>
