+++ app/views/exams/show.html.erb
@@ -9,16 +9,16 @@
         railsExam: {
           id: @exam.id,
           name: @exam.name,
-          policies: @exam.policies
+          policies: @exam.info['policies']
         },
         contents: {
           exam: {
-            questions: @exam.info[:contents][:questions],
-            reference: @exam.info[:contents][:reference],
-            instructions: @exam.info[:contents][:instructions],
-            files: @exam.get_exam_files
+            questions: @exam.info['contents']['questions'],
+            reference: @exam.info['contents']['reference'],
+            instructions: @exam.info['contents']['instructions'],
+            files: @exam.files
           },
-          answers: @exam.info[:answers]
+          answers: @exam.info['answers']
         }
       }) %>
     </div>
@@ -37,7 +37,7 @@
     railsExam: {
       id: @exam.id,
       name: @exam.name,
-      policies: @exam.policies
+      policies: @exam.info['policies']
     },
     railsRegistration: {
       id: @registration.id,
