--- db/migrate/20190513234634_create_uploads.rb
@@ -1,13 +1,0 @@
-class CreateUploads < ActiveRecord::Migration[5.2]
-  def change
-    create_table :uploads do |t|
-      t.integer :user_id, null: false
-      t.string :file_name, null: false
-      t.string :secret_key, null: false
-      t.integer :exam_id, null: false
-
-      t.index :exam_id, unique: true
-      t.timestamps
-    end
-  end
-end
