+++ db/migrate/20190513235327_create_exams.rb
@@ -5,6 +5,9 @@ class CreateExams < ActiveRecord::Migration[5.2]
       t.boolean :enabled, null: false
       t.string :name, null: false
 
+      t.jsonb :files, null: false
+      t.jsonb :info, null: false
+
       t.timestamps
     end
   end
