+++ db/schema.rb
@@ -35,6 +35,8 @@ ActiveRecord::Schema.define(version: 2020_05_14_210728) do
     t.string "secret_key", null: false
     t.boolean "enabled", null: false
     t.string "name", null: false
+    t.jsonb "files", null: false
+    t.jsonb "info", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
@@ -61,16 +63,6 @@ ActiveRecord::Schema.define(version: 2020_05_14_210728) do
     t.jsonb "answers", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
-  end
-
-  create_table "uploads", force: :cascade do |t|
-    t.integer "user_id", null: false
-    t.string "file_name", null: false
-    t.string "secret_key", null: false
-    t.integer "exam_id", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
-    t.index ["exam_id"], name: "index_uploads_on_exam_id", unique: true
   end
 
   create_table "users", force: :cascade do |t|
