+++ test/fixtures/exams.yml
@@ -1,9 +1,32 @@
+<%
+  # TODO abstract this into a test helper, and rescue, deleting zip
+  cs2500zip = Rails.root.join("test", "fixtures", "files", "cs2500midterm.zip")
+  ArchiveUtils.create_zip cs2500zip, Dir.glob(Rails.root.join("test", "fixtures", "files", "cs2500midterm", "**"))
+  cs2500upload = Upload.new(
+    FakeUpload.new(cs2500zip),
+    User.find(ActiveRecord::FixtureSet.identify(:ben))
+  )
+  FileUtils.remove_entry_secure cs2500zip
+%>
 cs2500midterm:
   secret_key: "94ae378e-53f3-4022-94cf-cbe6fe0c6a03"
   enabled: true
   name: "CS2500 Midterm"
+  info: <%= cs2500upload.info %>
+  files: <%= cs2500upload.files %>
 
+<%
+  cs3500zip = Rails.root.join("test", "fixtures", "files", "cs3500final.zip")
+  ArchiveUtils.create_zip cs3500zip, Dir.glob(Rails.root.join("test", "fixtures", "files", "cs3500final", "**"))
+  cs3500upload = Upload.new(
+    FakeUpload.new(cs3500zip),
+    User.find(ActiveRecord::FixtureSet.identify(:amit))
+  )
+  FileUtils.remove_entry_secure cs3500zip
+%>
 cs3500final:
   secret_key: "f772d1aa-4def-4af5-b650-d353519f3dad"
   enabled: true
   name: "CS3500 Final"
+  info: <%= cs3500upload.info %>
+  files: <%= cs3500upload.files %>
